+++ src/Database/ClientEntity.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Heptacom\AdminOpenAuth\Database;
 
+use Shopware\Core\Framework\Api\Acl\Role\AclRoleCollection;
 use Shopware\Core\Framework\DataAbstractionLayer\Entity;
 use Shopware\Core\Framework\DataAbstractionLayer\EntityIdTrait;
 
@@ -34,6 +35,14 @@ class ClientEntity extends Entity
     protected ?UserKeyCollection $userKeys = null;
 
     protected ?UserTokenCollection $userTokens = null;
+
+    protected AclRoleCollection $defaultAclRoles;
+
+    public function __construct()
+    {
+        $this->defaultAclRoles = new AclRoleCollection();
+    }
+
 
     public function getName(): ?string
     {
@@ -177,5 +186,15 @@ class ClientEntity extends Entity
         $this->userTokens = $userTokens;
 
         return $this;
+    }
+
+    public function getDefaultAclRoles(): AclRoleCollection
+    {
+        return $this->defaultAclRoles;
+    }
+
+    public function setDefaultAclRoles(AclRoleCollection $defaultAclRoles): void
+    {
+        $this->defaultAclRoles = $defaultAclRoles;
     }
 }