+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 
+* Added client configuration to assign default roles to new non admin users
 * Changed `views/administration/heptacom-admin-open-auth/page/confirm.html.twig` to immediately close the window after local storage item was set
 
 # 4.1.0